Fix warnings. Add simple docking menu in the sandbox.

+++ Razor/src/Platform/Windows/WindowsWindow.h
@@ -15,8 +15,8 @@ namespace Razor {
 
 		void OnUpdate() override;
 
-		inline unsigned int GetWidth() const override { return m_Data.Width; }
-		inline unsigned int GetHeight() const override { return m_Data.Height; }
+		inline float GetWidth() const override { return m_Data.Width; }
+		inline float GetHeight() const override { return m_Data.Height; }
 
 		// Window attributes
 		inline void SetEventCallback(const EventCallbackFn& callback) override { m_Data.EventCallback = callback; }
@@ -33,7 +33,7 @@ namespace Razor {
 		struct WindowData
 		{
 			std::string Title;
-			unsigned int Width, Height;
+			float Width, Height;
 			bool VSync;
 
 			EventCallbackFn EventCallback;
